# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AICar AutoCommerce API",
        "title": "AICar API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/cars": {
            "get": {
                "description": "키워드, 브랜드, 연료 타입, 가격 범위, 연도 범위로 차량을 검색합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "차량 검색/목록 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "검색어 (브랜드명, 모델명, 트림명)",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "페이지 번호 (기본값: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "페이지 크기 (기본값: 10)",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "브랜드 ID",
                        "name": "brand_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "연료 타입",
                        "name": "fuel_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "최소 가격",
                        "name": "min_price",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "최대 가격",
                        "name": "max_price",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "최소 연도",
                        "name": "min_year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "최대 연도",
                        "name": "max_year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "정렬 (price_asc, price_desc, year_asc, year_desc, created_at_asc, created_at_desc)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cars/{id}": {
            "get": {
                "description": "차량 ID로 상세 정보를 조회합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "차량 상세 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "차량 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cars/{id}/images": {
            "get": {
                "description": "차량 ID로 이미지 목록을 조회합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cars"
                ],
                "summary": "차량 이미지 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "차량 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}